update tests after utils changes
